Put drive-by times into their own component, add color to indicate how far from the queried time it is and make them clickable.
